TP3/UI/pyproject.toml
[tool.poetry]
name = "mips-pipeline-interface"
version = "0.1.0"
description = ""
authors = ["Bruno Guglielmotti <brunoguglielmotti@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
